# harness/runs/opencode-kimi-seed1/evidence/api-design/openapi.yaml
openapi: 3.1.0
info:
  title: Library Lending API
  version: 1.0.0
  description: API for managing library book lending, returns, reservations, and fines
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
  /books:
    get:
      summary: List all books
      operationId: listBooks
      responses:
        '200':
          description: List of books with availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Create a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookInput'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /books/{id}:
    get:
      summary: Get a single book
      operationId: getBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookWithQueue'
        '404':
          $ref: '#/components/responses/NotFound'
  /members:
    get:
      summary: List all members
      operationId: listMembers
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
    post:
      summary: Create a new member
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberInput'
      responses:
        '201':
          description: Member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /members/{id}:
    get:
      summary: Get member with stats
      operationId: getMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member details with stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberWithStats'
        '404':
          $ref: '#/components/responses/NotFound'
  /members/{id}/loans:
    get:
      summary: Get member's active loans
      operationId: getMemberLoans
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of active loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanWithBook'
  /members/{id}/reservations:
    get:
      summary: Get member's reservations
      operationId: getMemberReservations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationWithBook'
  /members/{id}/fines:
    get:
      summary: Get member's unpaid fines
      operationId: getMemberFines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of unpaid fines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fine'
  /loans:
    post:
      summary: Create a loan (checkout)
      operationId: createLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanInput'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Outstanding fines exceed limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Member is suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'
  /loans/{id}/return:
    post:
      summary: Return a book
      operationId: returnLoan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Loan already returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reservations:
    post:
      summary: Create a reservation
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationInput'
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /reservations/{id}/cancel:
    post:
      summary: Cancel a reservation
      operationId: cancelReservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /fines/{id}/pay:
    post:
      summary: Pay a fine
      operationId: payFine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fine paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fine'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Fine already paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - author
        - isbn
        - total_copies
        - available_copies
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        total_copies:
          type: integer
          minimum: 1
        available_copies:
          type: integer
          minimum: 0
    BookWithQueue:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            queue_depth:
              type: integer
              minimum: 0
    CreateBookInput:
      type: object
      required:
        - title
        - author
        - isbn
        - total_copies
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        total_copies:
          type: integer
          minimum: 1
    Member:
      type: object
      required:
        - id
        - name
        - email
        - member_since
        - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        member_since:
          type: string
          format: date
        status:
          type: string
          enum: [active, suspended]
    MemberWithStats:
      allOf:
        - $ref: '#/components/schemas/Member'
        - type: object
          properties:
            active_loans:
              type: integer
              minimum: 0
            unpaid_fines_cents:
              type: integer
              minimum: 0
    CreateMemberInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    Loan:
      type: object
      required:
        - id
        - book_id
        - member_id
        - borrowed_at
        - due_at
        - returned_at
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        borrowed_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
          nullable: true
    LoanWithBook:
      allOf:
        - $ref: '#/components/schemas/Loan'
        - type: object
          properties:
            book_title:
              type: string
            days_overdue:
              type: integer
              minimum: 0
    CreateLoanInput:
      type: object
      required:
        - member_id
        - book_id
      properties:
        member_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
    Reservation:
      type: object
      required:
        - id
        - book_id
        - member_id
        - queued_at
        - status
        - notified_at
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        queued_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [waiting, notified, expired, fulfilled, cancelled]
        notified_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    ReservationWithBook:
      allOf:
        - $ref: '#/components/schemas/Reservation'
        - type: object
          properties:
            book_title:
              type: string
            queue_position:
              type: integer
              minimum: 1
    CreateReservationInput:
      type: object
      required:
        - member_id
        - book_id
      properties:
        member_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
    Fine:
      type: object
      required:
        - id
        - member_id
        - loan_id
        - amount_cents
        - paid_at
        - book_title
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        loan_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
          minimum: 0
        paid_at:
          type: string
          format: date-time
          nullable: true
        book_title:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'